+++ BackEnd/ProjetoInt.WebApi/ProjetoInt.WebApi/Domains/TipoUsuario.cs
@@ -12,7 +12,7 @@ namespace ProjetoInt.WebApi.Domains
             Usuarios = new HashSet<Usuario>();
         }
 
-        public short IdTipoUsuario { get; set; }
+        public int IdTipoUsuario { get; set; }
         public string NomeTipoUsuario { get; set; }
 
         public virtual ICollection<Usuario> Usuarios { get; set; }
