+++ BackEnd/ProjetoInt.WebApi/ProjetoInt.WebApi/Domains/Usuario.cs
@@ -7,8 +7,8 @@ namespace ProjetoInt.WebApi.Domains
 {
     public partial class Usuario
     {
-        public short IdUsuario { get; set; }
-        public short? IdTipoUsuario { get; set; }
+        public int IdUsuario { get; set; }
+        public int? IdTipoUsuario { get; set; }
         public string Email { get; set; }
         public string Senha { get; set; }
         public string Nome { get; set; }
